moved out locale screenshare component to separate file
